+++ SourceCode/App/Pages/Country/CountryFilter.cs
@@ -1,0 +1,6 @@
+﻿namespace ModulesRegistry.Pages.Country;
+
+public record CountryFilter
+{
+    public int CountryId { get; set; }
+}