+++ RELEASENOTES.md
@@ -2,6 +2,10 @@
 > Release notes are only published in English.
 The release notes are summaries of important changes and fixes in each release. 
 There is maximum one release per date, but this can be performed as one or several deployments during a day.
+## Release 1.11.7
+Release date 2024-08-21
+- **Group members** can now be sent mails by group administrators.
+- **Vehicles administrator** page now displays all vehicles in the selected country.
 ## Release 1.11.6
 Release date 2024-08-20
 - **Group members** now have new right *is meeting administrator*.